Updated protocol, breaking compatibility
- increased SeqCount size to 32bit
- added flags
- added TTL support

+++ rest/data.go
@@ -17,7 +17,7 @@ type Contact struct {
 
 // CQMessage indicates the users location.
 type CQMessage struct {
-	Sequence uint16  `json:"sequence"`
+	Sequence uint32  `json:"sequence"`
 	Contact  Contact `json:"contact"`
 	Message  string  `json:"message"`
 }
